Added rate limit to makeExecutableSchema
